requirements.txt
discord.py==1.2.2
pyliner==1.0
sqlitehouse==1.2